requirements.txt
Flask~=2.2.2
text_to_owo~=2.0.1
